refactor: add builder pattern to existing services (#10)
* refactor: add builder pattern to existing services

* fix: Photo crud update & add builder pattern

* fix: remove @notBlank from person model

* refactor: change genderId type to int in PersonRequestDTO and GenderRepository

+++ src/main/java/com/la_copine/api/repository/GenderRepository.java
@@ -3,5 +3,5 @@ package com.la_copine.api.repository;
 import com.la_copine.api.model.Gender;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GenderRepository extends JpaRepository<Gender, Long> {
+public interface GenderRepository extends JpaRepository<Gender, Integer> {
 }
